BrianQC fixes
