+++ psi4/src/psi4/libfock/brianqc_int.h
@@ -29,6 +29,7 @@
 #ifndef LIBFOCK_BRIANQC_INT_H
 #define LIBFOCK_BRIANQC_INT_H
 #ifdef USING_BrianQC
+#include "cubature.h"
 #include "integrator_manager.h"
 
 namespace psi {
@@ -52,30 +53,32 @@ struct BrianBlock {
 class BrianQCBase : public IntegratorManager {
    using IntegratorManager::IntegratorManager;
     
-   void initialize() override;
    static void initialize_named_grids();
-   void grid_from_options(MolecularGridOptions const& opt, bool build_dft=true);
-   void grid_from_options(PruneSpec const& opt, int radscheme, std::vector<BrianBlock>& brian_blocks);
+   void grid_from_options(MolecularGrid::MolecularGridOptions const& opt, bool build_dft=true);
+   static void grid_from_prune_spec(StandardGridMgr::PruneSpec const& spec, int radscheme, std::vector<BrianBlock>& brian_blocks);
+
+   public:
+     void initialize() override;
 
    protected:
-    /// Integrator object for GauXC based integration
-    std::shared_ptr<GauXC::XCIntegrator<Eigen::MatrixXd>> integrator_;
     /// SG0 and SG1... It's an array of 4 for historical reasons.
     /// Array specifies which named grid. Outer vector specifies atom.
-    static std::array<std::vector<std::vector<BrianBlock>>, 4> brian_standard_grids;
+    static inline std::array<std::vector<std::vector<BrianBlock>>, 4> brian_standard_grids;
 
 };
 
 class BrianRV : public BrianQCBase {
     using BrianQCBase::BrianQCBase;
 
-    std::map<std::string, double> compute_V(std::vector<SharedMatrix> ret) override;
+    public:
+     std::map<std::string, double> compute_V(std::vector<SharedMatrix> ret) override;
 };
 
 class BrianUV : public BrianQCBase {
     using BrianQCBase::BrianQCBase;
 
-    std::map<std::string, double> compute_V(std::vector<SharedMatrix> ret) override;
+    public:
+     std::map<std::string, double> compute_V(std::vector<SharedMatrix> ret) override;
 };
 }
