+++ psi4/src/psi4/libfock/v.h
@@ -42,7 +42,12 @@ class DFTGrid;
 class PointFunctions;
 class SuperFunctional;
 class BlockOPoints;
+#ifdef USING_BrianQC
+class BrianQCBase;
+#endif
+#ifdef USING_gauxc
 class GauXCBase;
+#endif
 
 // => BASE CLASS <= //
